Erweiterung für den Controller static list für eine rückgabe Simulation und id based request

+++ Assessment/Models/TimelogTypeResponse.cs
@@ -4,6 +4,6 @@ namespace Assessment.Models
     {
         public int timelogTypeId { get; set; }
         public string timelogType { get; set; }
-        public int budget { get; set; }
+        public double budget { get; set; }
     }
 }
